# src/desafio-tecnico-sicredi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Desafio Técnico Sicredi",
        "description": "Aplicação de Votação",
        "contact": {
            "name": "Adriano Anschau",
            "email": "<adrianoanschau@gmail.com>"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://desafio-tecnico-sicredi.herokuapp.com/api/v1"
        }
    ],
    "paths": {
        "/api/v1/associates": {
            "get": {
                "tags": [
                    "Listar"
                ],
                "summary": "Listar Associados",
                "operationId": "associateList",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cadastrar"
                ],
                "summary": "Cadastrar um Associado",
                "operationId": "associateStore",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 3
                        }
                    },
                    {
                        "name": "document",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 11,
                            "minLength": 11
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Documento já cadastrado",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/associates/1": {
            "get": {
                "tags": [
                    "Exibir"
                ],
                "summary": "Exibir um Associado",
                "operationId": "associateShow",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Associado não Encontrado",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Atualizar"
                ],
                "summary": "Atualizar um Associado",
                "operationId": "associateUpdate",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "minLength": 3
                        }
                    },
                    {
                        "name": "document",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "maxLength": 11,
                            "minLength": 11
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Documento já cadastrado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Associado não Encontrado",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Excluir"
                ],
                "summary": "Excluir um Associado",
                "operationId": "associateDestroy",
                "responses": {
                    "204": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Associado não Encontrado",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    }
}